+++ aim/web/ui/src/components/ChartPanel/ChartPanelResizingFallback/index.ts
@@ -1,0 +1,3 @@
+import ChartPanelResizingFallback from './ChartPanelResizingFallback';
+
+export default ChartPanelResizingFallback;