+++ aim/web/ui/src/config/table/tableConfigs.ts
@@ -37,7 +37,7 @@ export const SELECTION_COLUMN_WIDTH = 32;
 
 export const VIEW_PORT_OFFSET = 500;
 
-export const TABLE_DEFAULT_CONFIG = {
+export const TABLE_DEFAULT_CONFIG: Record<string, any> = {
   [AppNameEnum.RUNS]: {
     resizeMode: ResizeModeEnum.Resizable,
     rowHeight: RowHeightSize.md,
@@ -127,21 +127,44 @@ export const AVOID_COLUMNS_TO_HIDE_LIST = new Set([
   'context',
 ]);
 
-export const EXPLORE_SELECTED_RUNS_CONFIG: {
-  [key: string]: AppNameEnum[];
-} = {
-  [AppNameEnum.RUNS]: [AppNameEnum.METRICS, AppNameEnum.IMAGES],
-  [AppNameEnum.METRICS]: [AppNameEnum.RUNS, AppNameEnum.IMAGES],
+export const EXPLORE_SELECTED_RUNS_CONFIG: Record<string, AppNameEnum[]> = {
+  [AppNameEnum.RUNS]: [
+    AppNameEnum.METRICS,
+    AppNameEnum.IMAGES,
+    AppNameEnum.FIGURES,
+    AppNameEnum.AUDIOS,
+  ],
+  [AppNameEnum.METRICS]: [
+    AppNameEnum.RUNS,
+    AppNameEnum.IMAGES,
+    AppNameEnum.FIGURES,
+    AppNameEnum.AUDIOS,
+  ],
   [AppNameEnum.PARAMS]: [
     AppNameEnum.RUNS,
     AppNameEnum.IMAGES,
     AppNameEnum.METRICS,
+    AppNameEnum.FIGURES,
+    AppNameEnum.AUDIOS,
   ],
   [AppNameEnum.SCATTERS]: [
     AppNameEnum.RUNS,
     AppNameEnum.IMAGES,
     AppNameEnum.METRICS,
+    AppNameEnum.FIGURES,
+    AppNameEnum.AUDIOS,
   ],
-  [AppNameEnum.IMAGES]: [AppNameEnum.RUNS, AppNameEnum.METRICS],
-  dashboard: [AppNameEnum.RUNS, AppNameEnum.METRICS, AppNameEnum.IMAGES],
+  [AppNameEnum.IMAGES]: [
+    AppNameEnum.RUNS,
+    AppNameEnum.METRICS,
+    AppNameEnum.FIGURES,
+    AppNameEnum.AUDIOS,
+  ],
+  dashboard: [
+    AppNameEnum.RUNS,
+    AppNameEnum.METRICS,
+    AppNameEnum.IMAGES,
+    AppNameEnum.FIGURES,
+    AppNameEnum.AUDIOS,
+  ],
 };