+++ aim/web/ui/src/types/services/models/metrics/metricsAppModel.d.ts
@@ -20,7 +20,12 @@ import {
   AggregationAreaMethods,
   AggregationLineMethods,
 } from 'utils/aggregateGroupData';
-import { AlignmentOptionsEnum, CurveEnum, ZoomEnum } from 'utils/d3';
+import {
+  AlignmentOptionsEnum,
+  CurveEnum,
+  ZoomEnum,
+  LegendsModeEnum,
+} from 'utils/d3';
 import { IRequestProgress } from 'utils/app/setRequestProgress';
 import { SmoothingAlgorithmEnum } from 'utils/smoothingData';
 
@@ -41,6 +46,7 @@ export interface IMetricAppModelState {
   lineChartData: ILine[][];
   chartTitleData: IChartTitleData;
   aggregatedData: IAggregatedData[];
+  legendsData: LegendsDataType;
   tooltip: ITooltip;
   tableData: any[];
   tableColumns: ITableColumn[];
@@ -106,10 +112,15 @@ export interface ITooltip extends Partial<ITooltipConfig> {
   content?: ITooltipContent;
 }
 
+export interface LegendsConfig {
+  display: boolean;
+  mode: LegendsModeEnum;
+}
+
 export interface IMetricsCollection<T> {
   key?: string;
   groupKey?: string;
-  config: { [key: string]: string } | null;
+  config: Record<string, unknown> | null;
   color: string | null;
   dasharray: string | null;
   chartIndex: number;
@@ -253,4 +264,15 @@ export interface ISmoothing {
   factor: number;
   curveInterpolation: CurveEnum;
   isApplied: boolean;
+}
+
+export interface LegendColumnDataType {
+  color?: string;
+  dasharray?: string;
+  chartIndex?: number;
+  value: string;
+}
+
+export interface LegendsDataType {
+  [key: string]: Record<string, LegendColumnDataType[]>;
 }