+++ CHANGELOG.md
@@ -4,7 +4,9 @@
 
 ### Enhancements:
 
+- Add chart legends to the Metrics Explorer (KaroMourad)
 - Implement vertically scalable version of Remote Tracking (mihran113, alberttorosyan)
+- Add the ability to search, filter, and compare audio through Audios Explorer (VkoHov)
 - Add epoch tracking for PyTorch Lightning (tmynn)
 - Add PaddlePaddle integration (tmynn)
 - Add Optuna integration (tmynn)